requirements.txt
mcp[cli]==1.6.0 
python-dotenv
ipykernel
httpx
